Remove css-loader peerDependency
Hey,
I know you haven't been on this project for a while, but I'm using it for a project of mine.  
Forking it as my own is also a problem because I need the npm registry.
Anyway, `css-loader` is not needed as peerDependency, absurd is a css\html templating engine.